Support yarn v2 or greater (#2)
Add support for yarn v2 or greater

This is an initial support to make the parser more useful, but going forward we would likely want to simplify the single parsers by detecting the yarn.lock version (no __metadata means v1, otherwise we have the version field in __metadata), and have specific parsers for e.g. version and dependencies (while in this PR we somehow juggle through the minor differences between the two formats).